rename test fixtures
