+++ packages/widget/test/fixtures/anotherDocmapWithOneStep.ts
