+++ packages/widget/test/fixtures/docmapWithMultipleSteps.ts
