+++ packages/widget/test/fixtures/docmapWithOneStep.ts
